pyproject.toml
[tool.poetry]
name = "numerical-collection-cpp"
version = "0.0.0"
description = ""
authors = ["Kenta Kabashima <kenta_program37@hotmail.co.jp>"]
readme = "README.md"
packages = [{ include = "scripts" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
conan = "^1.56.0"

[tool.poetry.group.dev.dependencies]
adr-tools-python = "^1.0.3"
black = "^22.12.0"
click = "^8.1.3"
cmakelang = "^0.6.13"
doc8 = "^0.11.2"
kaleido = "==0.2.1"
myst-parser = "^0.18.1"
pandas = "^1.5.2"
plotly = "^5.11.0"
pre-commit = "^2.21.0"
pylint = "^2.15.9"
sphinx = "^5.3.0"
sphinx-autobuild = "^2021.3.14"
sphinx-rtd-theme = "^1.1.1"
sphinxcontrib-plantuml = "^0.24.1"
tqdm = "^4.64.1"
lizard = "^1.17.10"
lcov-cobertura = "^2.0.2"
clang-tidy-checker = "^0.1.2"
pytest = "^7.2.0"
pytest-approvaltests = "^0.2.4"
fastjsonschema = "^2.16.2"
pytomlpp = "^1.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
